refactor custom exception
