+++ qanary_component-NER-DBpedia-Spotlight/src/main/java/eu/wdaqua/qanary/exceptions/DBpediaSpotlightServiceNotAvailable.java
@@ -1,4 +1,4 @@
-package eu.wdaqua.qanary.dbpediaSpotlight;
+package eu.wdaqua.qanary.exceptions;
 
 public class DBpediaSpotlightServiceNotAvailable extends Exception {
     public DBpediaSpotlightServiceNotAvailable(String message) {
